requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras
altair<5
openai
tiktoken
qdrant-client
huggingface-hub==0.14.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2